# apps/backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.Article:
    properties:
      author:
        type: string
      category:
        $ref: '#/definitions/models.Category'
      category_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      img:
        type: string
      is_active:
        type: boolean
      likes:
        type: integer
      meta_description:
        type: string
      meta_img:
        type: string
      meta_keywords:
        type: string
      published_at:
        type: string
      slug:
        type: string
      tenant_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      views:
        type: integer
    type: object
  models.Banner:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      img:
        type: string
      is_active:
        type: boolean
      link:
        type: string
      order:
        type: integer
      tenant_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      valid_from:
        type: string
      valid_until:
        type: string
    type: object
  models.Category:
    properties:
      children:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      created_at:
        type: string
      desc:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      order:
        type: integer
      parent:
        $ref: '#/definitions/models.Category'
      parent_id:
        type: string
      paymentMethods:
        items:
          $ref: '#/definitions/models.PaymentMethod'
        type: array
      slug:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  models.PaymentGateway:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      metadata:
        type: string
      name:
        type: string
      order:
        type: integer
      paymentMethods:
        items:
          $ref: '#/definitions/models.PaymentMethod'
        type: array
      updated_at:
        type: string
    type: object
  models.PaymentMethod:
    properties:
      additional_fields:
        description: could be JSON
        type: string
      category:
        $ref: '#/definitions/models.Category'
      category_id:
        type: string
      code:
        type: string
      created_at:
        type: string
      desc:
        type: string
      expiry_seconds:
        type: integer
      fee_fixed:
        type: number
      fee_percent:
        type: number
      id:
        type: string
      img:
        type: string
      invoice_code:
        type: string
      is_active:
        type: boolean
      is_international:
        type: boolean
      is_ready_production:
        type: boolean
      max_amount:
        type: number
      min_amount:
        type: number
      name:
        type: string
      order:
        type: integer
      payment_gateway_id:
        type: string
      paymentGateway:
        $ref: '#/definitions/models.PaymentGateway'
      ppn:
        type: number
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  responses.Pagination:
    properties:
      from:
        type: integer
      has_more:
        type: boolean
      page:
        type: integer
      page_size:
        type: integer
      to:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  utils.OrderedErrorResponse:
    properties:
      details: {}
      dev_message:
        type: string
      error_code:
        type: string
      errors: {}
      message:
        type: string
      success:
        type: boolean
      trace:
        type: string
    type: object
  utils.OrderedSuccessResponse:
    properties:
      data: {}
      message:
        type: string
      meta: {}
      pagination:
        $ref: '#/definitions/responses.Pagination'
      success:
        type: boolean
    type: object
host: localhost:9001
info:
  contact: {}
  description: Ini adalah dokumentasi API untuk backend berbasis Go Fiber.
  title: DilZ Backend API
  version: "1.0"
paths:
...
  /v1/activity-logs/{id}:
    get:
      consumes:
      - application/json
      description: Mengambil satu baris data riwayat aktivitas (Activity Log) berdasarkan
        ID.
      parameters:
      - description: ID dari Activity Log
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mendapatkan data activity log
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
        "404":
          description: Activity Log tidak ditemukan
          schema:
            $ref: '#/definitions/utils.OrderedErrorResponse'
        "500":
          description: Terjadi kesalahan pada server
          schema:
            $ref: '#/definitions/utils.OrderedErrorResponse'
      summary: Ambil detail Activity Log
      tags:
      - Activity Log
  /v1/articles:
    get:
      consumes:
      - application/json
      description: Mengambil daftar artikel yang sudah dipublikasikan (published)
        untuk blog CodeXtory. Mendukung limitasi jumlah data menggunakan query parameter.
      parameters:
      - default: 10
        description: Batasan jumlah artikel yang dikembalikan
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mendapatkan data article
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
        "500":
          description: Gagal mendapatkan data article
          schema:
            $ref: '#/definitions/utils.OrderedErrorResponse'
      summary: Ambil daftar artikel publik
      tags:
      - Articles
  /v1/articles/{slug}:
    get:
      consumes:
      - application/json
      description: Mengambil satu data artikel yang dipublikasikan berdasarkan slug-nya.
        Terdapat opsi untuk memuat artikel terkait (related articles).
      parameters:
      - description: Slug dari artikel
        in: path
        name: slug
        required: true
        type: string
      - default: false
        description: Set true untuk menyertakan artikel terkait
        in: query
        name: related
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mendapatkan data article
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
        "404":
          description: Artikel tidak ditemukan
          schema:
            $ref: '#/definitions/utils.OrderedErrorResponse'
      summary: Ambil detail artikel
      tags:
      - Articles
  /v1/articles/popular:
    get:
      consumes:
      - application/json
      description: Mengambil daftar artikel yang paling banyak dibaca atau berstatus
        populer. Bisa mengecualikan artikel tertentu menggunakan parameter 'except'.
      parameters:
      - default: 10
        description: Batasan jumlah artikel yang dikembalikan
        in: query
        name: limit
        type: integer
      - description: ID atau slug artikel yang ingin dikecualikan dari hasil
        in: query
        name: except
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mendapatkan data article
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
        "404":
          description: Artikel tidak ditemukan
          schema:
            $ref: '#/definitions/utils.OrderedErrorResponse'
      summary: Ambil artikel populer
      tags:
      - Articles
  /v1/articles/process:
    post:
      consumes:
      - application/json
      description: Memproses data payload artikel. Jika field ID pada payload kosong,
        maka akan melakukan Insert (Tambah Artikel). Jika field ID terisi, akan melakukan
        Update pada artikel tersebut.
      parameters:
      - description: Data JSON Artikel
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.Article'
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil menambahkan atau mengupdate artikel
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
        "400":
          description: Gagal memproses data artikel dari body request
          schema:
            $ref: '#/definitions/utils.OrderedErrorResponse'
        "500":
          description: Gagal memproses/menyimpan artikel ke database
          schema:
            $ref: '#/definitions/utils.OrderedErrorResponse'
      summary: Tambah atau Update Artikel
      tags:
      - Articles
...
  /v1/banners:
    get:
      consumes:
      - application/json
      description: Mengambil daftar banner yang berstatus aktif dan masih berada dalam
        masa berlaku (valid). Mendukung limitasi jumlah data.
      parameters:
      - default: 10
        description: Batasan jumlah banner yang dikembalikan
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mendapatkan data banner
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
        "500":
          description: Gagal mendapatkan data banner
          schema:
            $ref: '#/definitions/utils.OrderedErrorResponse'
      summary: Ambil daftar banner aktif
      tags:
      - Banners
  /v1/banners/process:
    post:
      consumes:
      - application/json
      description: Memproses data payload banner. Jika field ID pada payload kosong,
        maka akan melakukan Insert (Tambah Banner). Jika field ID terisi, akan melakukan
        Update pada banner tersebut.
      parameters:
      - description: Data JSON Banner
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.Banner'
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil menambahkan atau mengupdate banner
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
        "400":
          description: Gagal memproses data banner dari body request
          schema:
            $ref: '#/definitions/utils.OrderedErrorResponse'
        "500":
          description: Gagal memproses/menyimpan banner ke database
          schema:
            $ref: '#/definitions/utils.OrderedErrorResponse'
      summary: Tambah atau Update Banner
      tags:
      - Banners
  /v1/categories:
    get:
      consumes:
      - application/json
      description: Mengambil semua data kategori yang tersedia.
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mendapatkan data category
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
      summary: Ambil daftar kategori
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Membuat entri data kategori baru ke dalam database.
      parameters:
      - description: Data JSON kategori baru
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Berhasil menambahkan category
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
      summary: Tambah kategori baru
      tags:
      - Categories
  /v1/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus data kategori dari database berdasarkan ID.
      parameters:
      - description: ID kategori
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil menghapus category
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
      summary: Hapus kategori
      tags:
      - Categories
    get:
      consumes:
      - application/json
      description: Mengambil satu baris data kategori secara spesifik menggunakan
        parameter ID.
      parameters:
      - description: ID kategori
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mendapatkan data category
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
      summary: Ambil detail kategori
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Memodifikasi keseluruhan data kategori berdasarkan ID.
      parameters:
      - description: ID kategori
        in: path
        name: id
        required: true
        type: string
      - description: Data JSON pembaruan kategori
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengupdate category
          schema:
            $ref: '#/definitions/utils.OrderedSuccessResponse'
      summary: Update kategori
      tags:
      - Categories
...
swagger: "2.0"